chore: Create settings page and backend logic

+++ packages/server/src/resolvers/authentication/deleteUser.mutation.ts
@@ -8,6 +8,7 @@ export interface deleteUserArgs {
 async function deleteUser(parent: any, { email, password }: deleteUserArgs, context: Context) {
   try {
     console.info('Trying to delete a user', { email });
+    if (!context.userId) throw new Error();
     await deleteUserService(email, password, context);
     console.log('User deletion successful', { email });
     return {
